+++ CHANGES.md
@@ -20,6 +20,8 @@ Features
 - new command `time`: allows time the execution of the given commands
 - the `create` command now supports async mode (async=true)
 - print last_zxid in `session_info` as hex
+- `json_dupes_for_keys` now accepts a parameter first, to include the
+  original non duplicated znode, too.
 
 1.1.0 (2015-06-17)
 --------------------